some more fixes

+++ src/Arena.cpp
@@ -97,6 +97,6 @@ void start() {
 	destroy();
 }
 
-void registerBot(const char *name, COLOR color, const char *img, void (*updateFn)(float), void (*initFn)()) {
+void registerBot(const char *name, COLOR color, const char *img, void (*updateFn)(double), void (*initFn)()) {
 	botsData.push_back({ name, color, img, updateFn, initFn });
 }
